// packages/theme/src/sys-tokens/elevation/jooc-sys-elevation-tokens.types.ts
export type JoocSysElevationTokensType = {
  none: string;
  deepTeal700: JoocSysElevationTokenSizesType;
};

export type JoocSysElevationTokenSizesType = {
  xs: JoocSysElevationTokensDirectionType;
  s: JoocSysElevationTokensDirectionType;
  m: JoocSysElevationTokensDirectionType;
  l: JoocSysElevationTokensDirectionType;
  xl: JoocSysElevationTokensDirectionType;
};

export type JoocSysElevationTokensDirectionType = {
  top: string;
  right: string;
  bottom: string;
  left: string;
};